Extract dummyForm to own component, add route to form
